authhandler: Add support for 3-legged-OAuth
Added authhandler.go, which implements a TokenSource
to support "three-legged OAuth 2.0" via a custom
AuthorizationHandler.

Added default_authhandler.go to provide a command line
implementation for AuthorizationHandler.
